# perf: saving SendToReplicas results with a bitmap