@PingXie 2025-07-07
no concerns. I think this is a good change and the smaller bitmap helps conserve the cache lines too. thanks @rueian.